only export MongoDB dependency for shared mongodb-library
libmongoclient.a uses quite a number of other libs and the exact
requirements can't be read from a cmake/pc file.
Therefore it makes more sense to keep the dependency hidden from ROS
when we use the static lib. libwarehouse_ros then provides all required functions.
... This is a bit like creating a libmongoclient.so, but the whole problem
exists because debian/ubuntu don't provide this one, right?

The shared library can - and has to - be exported as a dependency to ROS.