Fix public OpenSSL dependency
warehouse_ros is using `MD5()` from OpenSSL in inline code in a header,
so every consumer of the headers have to link against libcrypto.
By using the OpenSSL::Crypto target, ament can do that for us.
Due to the different cases of the OpenSSL module and variable names,
`ament_target_dependencies` does not work with plain old variables.